[math] Move ArithmeticTraits to :math:utils module
