+++ src/modm/architecture/driver/heap/block_allocator.hpp
@@ -19,7 +19,7 @@
 #include <cstddef>
 
 #include <modm/architecture/utils.hpp>
-#include <modm/utils/arithmetic_traits.hpp>
+#include <modm/math/utils/arithmetic_traits.hpp>
 
 namespace modm
 {
