+++ src/modm/io/module.lb
@@ -18,8 +18,7 @@ def init(module):
 def prepare(module, options):
     module.depends(
         ":architecture:accessor",
-        ":math:utils",
-        ":utils")
+        ":math:utils")
     return True
 
 def build(env):
