+++ src/modm/math/filter/module.lb
@@ -18,8 +18,7 @@ def init(module):
 def prepare(module, options):
     module.depends(
         ":architecture",
-        ":math:utils",
-        ":utils")
+        ":math:utils")
     return True
 
 def build(env):
