+++ src/modm/math/filter/pid.hpp
@@ -19,7 +19,7 @@
 #include <cmath>
 #include <stdint.h>
 
-#include <modm/utils/arithmetic_traits.hpp>
+#include <modm/math/utils/arithmetic_traits.hpp>
 
 namespace modm
 {
