+++ src/modm/processing/timer/timestamp.hpp
@@ -20,7 +20,7 @@
 #include <stdint.h>
 
 #include <modm/io/iostream.hpp>
-#include <modm/utils/arithmetic_traits.hpp>
+#include <modm/math/utils/arithmetic_traits.hpp>
 
 namespace modm
 {
