+++ src/modm/ui/animation/module.lb
@@ -20,7 +20,7 @@ Various classes for animating values.
 """
 
 def prepare(module, options):
-    module.depends(":architecture:clock", ":utils")
+    module.depends(":architecture:clock", ":math:utils")
     return True
 
 def build(env):
