+++ src/modm/utils/utils.hpp
@@ -15,5 +15,4 @@
 #include <modm/architecture/utils.hpp>
 
 #include "utils/allocator.hpp"
-#include "utils/arithmetic_traits.hpp"
 #include "utils/dummy.hpp"
