+++ test/config/nucleo-f103.xml
@@ -13,6 +13,5 @@
     <module>modm-test:test:processing</module>
     <module>modm-test:test:ui</module>
     <module>modm-test:test:math</module>
-    <module>modm-test:test:utils</module>
   </modules>
 </library>
