--- test/modm/utils/module.lb
@@ -1,30 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-#
-# Copyright (c) 2018, Christopher Durand
-#
-# This file is part of the modm project.
-#
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-
-def init(module):
-    module.name = "utils"
-    module.parent = "test"
-
-
-def prepare(module, options):
-    module.depends(":utils")
-
-    target = options[":target"].identifier
-    if target["platform"] == "avr":
-        module.depends(":stdc++")
-
-    return True
-
-
-def build(env):
-    env.outbasepath = "modm-test/src/modm/utils"
-    env.copy('.')
